not delete test

+++ app/Http/Controllers/UserController.php
@@ -50,7 +50,13 @@ class UserController extends Controller
     public function deleteAccount(Request $request)
     {
         $user = $request->user();
-        
+
+        if (in_array($user->phone, ['900000000', '900038902'])) {
+            return response()->json([
+                'message' => 'Account deleted successfully. Your data has been preserved for analytics purposes..'
+            ]);
+        }
+
         // Soft delete the user account
         $user->delete();
